Create Activity - Vlees
Vlees is just a random name that was left because I didn't want to completely restart this project again. This activity creates a launch screen, which ends after 3 seconds. It can be used in the final app, but was mainly used to show that the MainActivity wouldn't pop up again after first launch.